task : move special operators constants to internal package
