+++ pkg/internal/specops/spec_ops.go
@@ -1,4 +1,4 @@
-package defs
+package specops
 
 // ListActionsSpecOpFailedActionsLabel indicates that listActions should return only actions with status 'failed'.
 const ListActionsSpecOpFailedActionsLabel = "97d4eb1e49215e3374cc2c1939a7c43a55e95c7427bf2d45ed63e3b4e0c88153"
