+++ pkg/storage/internal/actions/list_actions_mapping.go
@@ -7,8 +7,8 @@ import (
 	"slices"
 
 	"github.com/bsv-blockchain/go-sdk/transaction"
-	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/defs"
 	pkgentity "github.com/bsv-blockchain/go-wallet-toolbox/pkg/entity"
+	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/specops"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/storage/entity"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/wdk"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/wdk/primitives"
@@ -25,7 +25,7 @@ func (l *listActions) toFilterParams(userID int, args *wdk.ListActionsArgs) (ent
 	isFailedQuery := false
 	filteredLabels := make([]string, 0, len(labelNames))
 	for _, label := range labelNames {
-		if defs.IsListActionsSpecOp(label) {
+		if specops.IsListActionsSpecOp(label) {
 			isFailedQuery = true
 			continue
 		}
